$ Compiling Successful

+++ include/RotationMatrix3by3.hpp
@@ -11,6 +11,6 @@ class RotationMatrix3by3 : public Matrix3by3
     public:
         RotationMatrix3by3(/* args */);
         ~RotationMatrix3by3();
-        void Update(Vector3D<float>);
+        void Update(Vector3D<double>);
 };
